Making dynamic inventory script for Vagrant work for Windows
When I was going through the examples on Windows, the `--list` option would not report any hosts even through I had two vagrant hosts running.  The problem was that the `vagrant status --machine-readable` command had carriage returns.  The last column was expected to be `'running'` but it in reality it was `'running\r'`.  I changed the script from using `.strip('\n')` to `.splitlines()` which not only handles carriage returns but discards the last _empty line_.